actually remove references to astar, should we rename timedep to astar?
