+++ test/multipoint_routes.cc
@@ -6,11 +6,9 @@
 
 #include "baldr/rapidjson_utils.h"
 #include "loki/worker.h"
-#include "midgard/logging.h"
 #include "midgard/util.h"
 #include "odin/worker.h"
 #include "sif/autocost.h"
-#include "thor/astar.h"
 #include "thor/worker.h"
 #include <boost/property_tree/ptree.hpp>
