+++ valhalla/thor/timedep.h
@@ -1,8 +1,22 @@
-#ifndef VALHALLA_THOR_TIMEDEP_H_
-#define VALHALLA_THOR_TIMEDEP_H_
+#pragma once
 
+#include <cstdint>
+#include <memory>
+#include <unordered_map>
+#include <utility>
+#include <vector>
+
+#include <valhalla/baldr/double_bucket_queue.h>
+#include <valhalla/baldr/graphid.h>
+#include <valhalla/baldr/graphreader.h>
 #include <valhalla/baldr/time_info.h>
-#include <valhalla/thor/astar.h>
+#include <valhalla/sif/dynamiccost.h>
+#include <valhalla/sif/edgelabel.h>
+#include <valhalla/sif/hierarchylimits.h>
+#include <valhalla/thor/astarheuristic.h>
+#include <valhalla/thor/edgestatus.h>
+#include <valhalla/thor/pathalgorithm.h>
+#include <valhalla/thor/pathinfo.h>
 
 namespace valhalla {
 namespace thor {
@@ -13,7 +27,7 @@ namespace thor {
  * For driving routes it uses a highway hierarchy with shortcut edges to
  * improve performance.
  */
-class TimeDepForward : public AStarPathAlgorithm {
+class TimeDepForward : public PathAlgorithm {
 public:
   /**
    * Constructor.
@@ -45,6 +59,11 @@ public:
               const Options& options = Options::default_instance()) override;
 
   /**
+   * Clear the temporary information generated during path construction.
+   */
+  virtual void Clear() override;
+
+  /**
    * Returns the name of the algorithm
    * @return the name of the algorithm
    */
@@ -52,7 +71,23 @@ public:
     return "time_dependent_forward_a*";
   }
 
+  /**
+   * Set a maximum label count. The path algorithm terminates if this
+   * is exceeded.
+   * @param  max_count  Maximum number of labels to allow.
+   */
+  void set_max_label_count(const uint32_t max_count) {
+    max_label_count_ = max_count;
+  }
+
 protected:
+  /**
+   * Initializes the hierarchy limits, A* heuristic, and adjacency list.
+   * @param  origll  Lat,lng of the origin.
+   * @param  destll  Lat,lng of the destination.
+   */
+  virtual void Init(const midgard::PointLL& origll, const midgard::PointLL& destll);
+
   /**
    * Expand from the node along the forward search path. Immediately expands
    * from the end node of any transition edge (so no transition edges are added
@@ -88,6 +123,71 @@ protected:
                                  const baldr::TimeInfo& time_info,
                                  const valhalla::Location& destination,
                                  std::pair<int32_t, float>& best_path);
+
+  /**
+   * Modify hierarchy limits based on distance between origin and destination
+   * and the relative road density at the destination. For shorter routes
+   * we stay on arterial roads further from the destination. Also for lower
+   * road densities near the destination the hierarchy transition distances
+   * are increased.
+   * @param   dist     Distance between origin and destination.
+   * @param   density  Relative road density near the destination.
+   */
+  void ModifyHierarchyLimits(const float dist, const uint32_t density);
+
+  /**
+   * Add edges at the origin to the adjacency list.
+   * @param  graphreader  Graph tile reader.
+   * @param  origin       Location information of the origin.
+   * @param  dest         Location information of the destination.
+   */
+  virtual void SetOrigin(baldr::GraphReader& graphreader,
+                         const valhalla::Location& origin,
+                         const valhalla::Location& destination,
+                         const uint32_t seconds_of_week);
+
+  /**
+   * Set the destination edge(s).
+   * @param   graphreader  Graph tile reader.
+   * @param   dest         Location information of the destination.
+   * @return  Returns the relative density near the destination (0-15)
+   */
+  virtual uint32_t SetDestination(baldr::GraphReader& graphreader, const valhalla::Location& dest);
+
+  /**
+   * Form the path from the adjacency list. Recovers the path from the
+   * destination backwards towards the origin (using predecessor information)
+   * @param   dest  Index in the edge labels of the destination edge.
+   * @return  Returns the path info, a list of GraphIds representing the
+   *          directed edges along the path - ordered from origin to
+   *          destination - along with travel modes and elapsed time.
+   */
+  virtual std::vector<PathInfo> FormPath(const uint32_t dest);
+
+  uint32_t max_label_count_; // Max label count to allow
+  sif::TravelMode mode_;     // Current travel mode
+  uint8_t travel_type_;      // Current travel type
+
+  // Hierarchy limits.
+  std::vector<sif::HierarchyLimits> hierarchy_limits_;
+
+  // A* heuristic
+  AStarHeuristic astarheuristic_;
+
+  // Current costing mode
+  std::shared_ptr<sif::DynamicCost> costing_;
+
+  // Vector of edge labels (requires access by index).
+  std::vector<sif::EdgeLabel> edgelabels_;
+
+  // Adjacency list - approximate double bucket sort
+  std::shared_ptr<baldr::DoubleBucketQueue> adjacencylist_;
+
+  // Edge status. Mark edges that are in adjacency list or settled.
+  EdgeStatus edgestatus_;
+
+  // Destinations, id and percent used along the edge
+  std::unordered_map<uint64_t, float> destinations_percent_along_;
 };
 
 /**
@@ -96,7 +196,7 @@ protected:
  * routes it uses a highway hierarchy with shortcut edges to improve
  * performance.
  */
-class TimeDepReverse : public AStarPathAlgorithm {
+class TimeDepReverse : public TimeDepForward {
 public:
   /**
    * Constructor.
@@ -231,5 +331,3 @@ protected:
 
 } // namespace thor
 } // namespace valhalla
-
-#endif // VALHALLA_THOR_TIMEDEP_H_
